INT-3496: revert `int` methods and introduce `long` methods for `count` metrics

+++ spring-integration-jmx/src/test/java/org/springframework/integration/monitor/HandlerMonitoringIntegrationTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2014 the original author or authors.
+ * Copyright 2009-2010 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -10,7 +10,6 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-
 package org.springframework.integration.monitor;
 
 import static org.junit.Assert.assertEquals;
@@ -85,7 +84,7 @@ public class HandlerMonitoringIntegrationTests {
 			channel.send(new GenericMessage<String>("bar"));
 			assertEquals(before + 1, service.getCounter());
 
-			long count = messageHandlersMonitor.getHandlerDuration(monitor).getCount();
+			int count = messageHandlersMonitor.getHandlerDuration(monitor).getCount();
 			assertTrue("No statistics for input channel", count > 0);
 
 		} finally {
